beginning of api refactor to endpiont based config away from classical while trying to debug caching
