+++ frontend/.env.production
@@ -14,4 +14,7 @@ SENDY_API_KEY=
 SENDY_API_URL=
 SENDY_LIST_ID=
 
-API_URL=http://api.simpler.grants.gov
+# API_URL=http://api.simpler.grants.gov
+
+API_URL=http://localhost:8080
+API_AUTH_TOKEN=LOCAL_AUTH_12345678
